--- main.ts
@@ -1,15 +1,0 @@
-import { Construct } from 'constructs';
-import { App, Chart, ChartProps } from 'cdk8s';
-
-export class MyChart extends Chart {
-  constructor(scope: Construct, id: string, props: ChartProps = { }) {
-    super(scope, id, props);
-
-    // define resources here
-
-  }
-}
-
-const app = new App();
-new MyChart(app, 'baseline-environment-on-kubernetes-cdk8s-typescript');
-app.synth();
